pyproject.toml
[project.scripts]
dataqe-run = "dataqe_framework.cli:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "dataqe-framework"
version = "0.1.0"
description = "Reusable Data Validation Framework"
authors = [
  { name="Khadar Shaik", email="khadarmohiddin.shaik@apree.health" }
]
dependencies = [
  "google-cloud-bigquery",
  "pymysql",
  "pyyaml",
  "pandas"
]
requires-python = ">=3.9"

